Create workshops.md
This is a simple placeholder to describe the Net Zero workshops and allow charges that are using the toolkit to register their interest. This process may change depending on the roll-out of the Net Zero workshop. But, first stage, let charges know they are coming and that they can register for these here.